refactor: 30% conversion to TS
