+++ gatsby-node.ts
@@ -3,7 +3,7 @@ const path = require('path');
 // Create pages from markdown files
 exports.createPages = ({ graphql, actions }) => {
  const { createPage } = actions;
- return new Promise((resolve, reject) => {
+ return new Promise<void>((resolve, reject) => {
   resolve(
    graphql(
     `
