+++ package.json
@@ -47,6 +47,7 @@
     "@iconify/icons-cib": "^1.2.4",
     "@iconify/icons-logos": "^1.2.36",
     "@iconify/react": "^4.1.1",
+    "@types/react-helmet": "^6.1.11",
     "babel-eslint": "^10.1.0",
     "eslint": "^8.55.0",
     "eslint-config-airbnb": "^19.0.4",
