+++ src/templates/service.tsx
@@ -1,14 +1,26 @@
 import React from 'react';
-import { graphql } from 'gatsby';
+import { graphql, PageProps } from 'gatsby';
 import SEO from '../components/SEO';
 import Layout from '../layouts/index';
 
-const Service = ({ data }) => {
+interface ServiceProps extends PageProps {
+  data: {
+    markdownRemark: {
+      frontmatter: {
+        title: string;
+      };
+      html: string;
+    };
+  };
+}
+
+const Service: React.FC<ServiceProps> = ({ data }) => {
   const { title } = data.markdownRemark.frontmatter;
   const { html } = data.markdownRemark;
+
   return (
     <Layout bodyClass="page-service">
-      <SEO title={title} />
+      <SEO title={title} meta={[]} keywords={[]} />
       <div className="strip strip-white strip-diagonal">
         <div className="container pt-4 pt-md-10">
           <div className="row justify-content-start">
