+++ src/templates/team.tsx
@@ -1,18 +1,29 @@
 import React from 'react';
-import { graphql } from 'gatsby';
+import { graphql, PageProps } from 'gatsby';
 import Layout from '../layouts/index';
 
-const Testimonial = ({ data }) => {
-  const { markdownRemark } = data;
-  const title = markdownRemark.frontmatter.title;
-  const html = markdownRemark.html;
+interface ITeam extends PageProps {
+  data: {
+    markdownRemark: {
+      frontmatter: {
+        title: string;
+      };
+      html: string;
+    };
+  };
+}
+
+const Team: React.FC<ITeam> = ({ data }) => {
+  const { title } = data.markdownRemark.frontmatter;
+  const { html } = data.markdownRemark;
+
   return (
-    <Layout bodyClass="page-testimonial">
+    <Layout bodyClass="page-team">
       <div className="strip strip-white strip-diagonal">
         <div className="container pt-4 pt-md-10">
           <div className="row justify-content-start">
             <div className="col-12 col-md-8">
-              <div className="testimonial testimonial-single">
+              <div className="team">
                 <h1 className="title">{title}</h1>
                 <div className="content" dangerouslySetInnerHTML={{ __html: html }} />
               </div>
@@ -36,4 +47,4 @@ export const query = graphql`
   }
 `;
 
-export default Testimonial;
+export default Team;
