+++ src/templates/testimonial.tsx
@@ -1,17 +1,29 @@
 import React from 'react';
-import { graphql } from 'gatsby';
+import { graphql, PageProps } from 'gatsby';
 import Layout from '../layouts/index';
 
-const Team = ({ data }) => {
-  const { title } = data.markdownRemark.frontmatter;
-  const { html } = data.markdownRemark;
+interface ITestimonial extends PageProps {
+  data: {
+    markdownRemark: {
+      frontmatter: {
+        title: string
+      }
+      html: string
+    };
+  }
+}
+
+const Testimonial: React.FC<ITestimonial> = ({ data }) => {
+  const { markdownRemark } = data;
+  const title = markdownRemark.frontmatter.title;
+  const html = markdownRemark.html;
   return (
-    <Layout bodyClass="page-team">
+    <Layout bodyClass="page-testimonial">
       <div className="strip strip-white strip-diagonal">
         <div className="container pt-4 pt-md-10">
           <div className="row justify-content-start">
             <div className="col-12 col-md-8">
-              <div className="team">
+              <div className="testimonial testimonial-single">
                 <h1 className="title">{title}</h1>
                 <div className="content" dangerouslySetInnerHTML={{ __html: html }} />
               </div>
@@ -35,4 +47,4 @@ export const query = graphql`
   }
 `;
 
-export default Team;
+export default Testimonial;
